+++ packages/kit/src/views/Onboarding/screens/ImportWallet/index.tsx
@@ -2,24 +2,9 @@ import { useCallback, useState } from 'react';
 
 import { useNavigation } from '@react-navigation/core';
 import { useRoute } from '@react-navigation/native';
-import { LinearGradient } from 'expo-linear-gradient';
 import { useIntl } from 'react-intl';
-import { StyleSheet } from 'react-native';
 
-import {
-  Box,
-  Center,
-  Icon,
-  Image,
-  Pressable,
-  Spinner,
-  Text,
-  useTheme,
-  useThemeValue,
-} from '@onekeyhq/components';
-import type { IconProps } from '@onekeyhq/components/src/Icon';
-import KeyTagPNG from '@onekeyhq/kit/assets/onboarding/import_with_keytag.png';
-import OneKeyLitePNG from '@onekeyhq/kit/assets/onboarding/import_with_lite.png';
+import { Box } from '@onekeyhq/components';
 import supportedNFC from '@onekeyhq/shared/src/detector/nfc';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
 
@@ -35,6 +20,14 @@ import Layout from '../../Layout';
 import { useOnboardingContext } from '../../OnboardingContext';
 import { EOnboardingRoutes } from '../../routes/enums';
 
+import {
+  OptionKeyTag,
+  OptionMigration,
+  OptionOneKeyLite,
+  OptionRecoveryPhrase,
+  OptioniCloud,
+} from './ImportWalletOptions';
+
 import type { IOnboardingRoutesParams } from '../../routes/types';
 import type { RouteProp } from '@react-navigation/native';
 import type { StackNavigationProp } from '@react-navigation/stack';
@@ -49,125 +42,13 @@ type NavigationProps = StackNavigationProp<
   EOnboardingRoutes.ImportWallet
 >;
 
-const ImportItem = ({
-  icon,
-  title,
-  children,
-  isDisabled,
-  isLoading,
-  onPress,
-}: {
-  icon: IconProps['name'];
-  title: string;
-  children: React.ReactNode;
-  isDisabled?: boolean;
-  isLoading?: boolean;
-  onPress: () => void;
-}) => {
-  const { themeVariant } = useTheme();
-
-  return (
-    <Box p="8px" w={{ base: '100%', sm: '1/3' }}>
-      <Pressable
-        flex={1} // Implement equal height on the same row
-        onPress={onPress}
-        p="16px"
-        pb="80px"
-        bg="action-secondary-default"
-        _hover={{ bgColor: 'action-secondary-hovered' }}
-        _pressed={{ bgColor: 'action-secondary-pressed' }}
-        borderWidth={StyleSheet.hairlineWidth}
-        borderColor="border-default"
-        borderRadius="12px"
-        overflow="hidden"
-        disabled={isDisabled}
-      >
-        <Box
-          flexDirection="row"
-          alignItems="center"
-          justifyContent="space-between"
-        >
-          <Icon
-            name={icon}
-            color={
-              isDisabled || isLoading ? 'icon-disabled' : 'interactive-default'
-            }
-          />
-          {isLoading && <Spinner />}
-        </Box>
-        <Text
-          mt={4}
-          typography={{ sm: 'Heading', md: 'Heading' }}
-          flex={1}
-          color={isDisabled || isLoading ? 'text-disabled' : 'text-default'}
-        >
-          {title}
-        </Text>
-        <Box
-          position="absolute"
-          bottom="-64px"
-          left="16px"
-          right="16px"
-          h="128px"
-          borderRadius="12px"
-          bg="surface-default"
-          overflow="hidden"
-        >
-          <LinearGradient
-            colors={['rgb(255, 255, 255)', 'rgba(255, 255, 255, 0)']}
-            style={{
-              position: 'absolute',
-              height: 64,
-              width: '100%',
-              opacity: 0.1,
-              zIndex: -1,
-            }}
-          />
-          {children}
-          <LinearGradient
-            colors={[
-              themeVariant === 'light'
-                ? 'rgba(255, 255, 255, 0)'
-                : 'rgba(30, 30, 42, 0)',
-              useThemeValue('surface-default'),
-            ]}
-            style={{
-              position: 'absolute',
-              height: '100%',
-              width: '100%',
-              bottom: 64,
-            }}
-          />
-        </Box>
-      </Pressable>
-    </Box>
-  );
-};
-
-const IconToIconIllus = ({
-  leftIcon,
-  rightIcon,
-}: {
-  leftIcon: IconProps['name'];
-  rightIcon: IconProps['name'];
-}) => (
-  <Center h="64px" flexDirection="row" px="24px">
-    <Icon name={leftIcon} size={32} color="icon-default" />
-    <LinearGradient
-      colors={['transparent', useThemeValue('icon-subdued'), 'transparent']}
-      start={{ x: 0, y: 0 }}
-      end={{ x: 1, y: 0 }}
-      style={{
-        flex: 1,
-        height: 2,
-        marginHorizontal: 8,
-      }}
-    />
-    <Icon name={rightIcon} size={32} color="icon-default" />
-  </Center>
-);
-
 const defaultProps = {} as const;
+
+const ItemWrapper = ({ children }: { children: any }) => (
+  <Box p="8px" w={{ base: '100%', sm: '1/3' }}>
+    {children}
+  </Box>
+);
 
 const ImportWallet = () => {
   const intl = useIntl();
@@ -218,8 +99,6 @@ const ImportWallet = () => {
     navigation.navigate(EOnboardingRoutes.RestoreFromCloud);
   }, [navigation]);
 
-  const FAKEPHRASE = ['shy', 'owner', 'almost', 'explain', 'movie', 'subway'];
-
   return (
     <Layout
       disableAnimation={disableAnimation}
@@ -227,73 +106,49 @@ const ImportWallet = () => {
       description={intl.formatMessage({ id: 'onboarding__import_wallet_desc' })}
     >
       <Box m={-2} flexDirection={{ base: 'column', sm: 'row' }} flexWrap="wrap">
-        <ImportItem
-          title={intl.formatMessage({
-            id: 'onboarding__import_wallet_with_recovery_phrase',
-          })}
-          icon="ClipboardDocumentListOutline"
-          onPress={onPressRecoveryWallet}
-        >
-          <Box flexDirection="row" px="16px" h="64px" flexWrap="wrap" py="8px">
-            {FAKEPHRASE.map((phrase, index) => (
-              <Box flexDirection="row" w="1/3" py="4px" key={index}>
-                <Text typography="Caption" w="12px" color="text-subdued">
-                  {index + 1}
-                </Text>
-                <Text typography="Caption">{phrase}</Text>
-              </Box>
-            ))}
-          </Box>
-        </ImportItem>
-        {supportedNFC && (
-          <ImportItem
+        <ItemWrapper>
+          <OptionRecoveryPhrase
             title={intl.formatMessage({
-              id: 'onboarding__import_wallet_with_lite',
+              id: 'onboarding__import_wallet_with_recovery_phrase',
             })}
-            icon="OnekeyLiteOutline"
-            onPress={onPressOneKeyLite}
-          >
-            <Center>
-              <Image source={OneKeyLitePNG} w="224px" h="64px" />
-            </Center>
-          </ImportItem>
-        )}
-        <ImportItem
-          title={intl.formatMessage({
-            id: 'onboarding__import_wallet_with_keytag',
-          })}
-          icon="KeytagOutline"
-          onPress={onPressKeyTag}
-        >
-          <Center>
-            <Image source={KeyTagPNG} w="224px" h="64px" />
-          </Center>
-        </ImportItem>
-        <ImportItem
-          title={intl.formatMessage({
-            id: 'onboarding__import_wallet_with_migrate',
-          })}
-          icon="ArrowPathRoundedSquareOutline"
-          onPress={onPressMigration}
-        >
-          <IconToIconIllus
-            leftIcon="ComputerDesktopSolid"
-            rightIcon="DevicePhoneMobileSolid"
+            onPress={onPressRecoveryWallet}
           />
-        </ImportItem>
-        {(platformEnv.isNativeIOS || platformEnv.isNativeIOSPad) && (
-          <ImportItem
-            title={intl.formatMessage({ id: 'action__restore_from_icloud' })}
-            icon="CloudOutline"
-            onPress={onPressRestoreFromCloud}
-            isDisabled={!hasPreviousBackups}
-            isLoading={iCloudLoading}
-          >
-            <IconToIconIllus
-              leftIcon="CloudSolid"
-              rightIcon="OnekeyLogoSolid"
+        </ItemWrapper>
+        {supportedNFC && (
+          <ItemWrapper>
+            <OptionOneKeyLite
+              title={intl.formatMessage({
+                id: 'onboarding__import_wallet_with_lite',
+              })}
+              onPress={onPressOneKeyLite}
             />
-          </ImportItem>
+          </ItemWrapper>
+        )}
+        <ItemWrapper>
+          <OptionKeyTag
+            title={intl.formatMessage({
+              id: 'onboarding__import_wallet_with_keytag',
+            })}
+            onPress={onPressKeyTag}
+          />
+        </ItemWrapper>
+        <ItemWrapper>
+          <OptionMigration
+            title={intl.formatMessage({
+              id: 'onboarding__import_wallet_with_migrate',
+            })}
+            onPress={onPressMigration}
+          />
+        </ItemWrapper>
+        {(platformEnv.isNativeIOS || platformEnv.isNativeIOSPad) && (
+          <ItemWrapper>
+            <OptioniCloud
+              title={intl.formatMessage({ id: 'action__restore_from_icloud' })}
+              onPress={onPressRestoreFromCloud}
+              isDisabled={!hasPreviousBackups}
+              isLoading={iCloudLoading}
+            />
+          </ItemWrapper>
         )}
       </Box>
     </Layout>
